requirements.txt
flask-smorest
python-dotenv
